Move connection code into separate hook.
